twitch_service/requirements_windows.txt
fastapi
uvicorn
redis
fastapi-cache2
sqlalchemy
asyncpg
pydantic
python-dotenv
pydantic_settings
